+++ service/src/bloom_filter.rs
@@ -1,7 +1,7 @@
-use crate::log::*;
-
 use std::collections::hash_map::DefaultHasher;
 use std::hash::Hasher;
+
+use crate::log::*;
 
 pub fn bloom_filter_bits(
   size: usize,
