+++ service/src/log.rs
@@ -1,11 +1,12 @@
+pub use std::sync::atomic::Ordering;
+use std::{sync::atomic::AtomicBool, sync::Mutex, thread};
+
 pub use crate::log_command;
 pub use crate::log_error;
 pub use crate::log_info;
 pub use crate::log_trace;
 pub use crate::log_verbose;
 pub use crate::log_warning;
-pub use std::sync::atomic::Ordering;
-use std::{sync::atomic::AtomicBool, sync::Mutex, thread};
 
 pub static ERROR: AtomicBool = AtomicBool::new(true);
 pub static WARNING: AtomicBool = AtomicBool::new(true);
