+++ service/src/main.rs
@@ -14,6 +14,7 @@ pub mod subgraph;
 pub mod vsids;
 pub mod write_ops;
 pub mod zero_opinion;
+pub mod settings;
 
 #[cfg(test)]
 mod tests;
@@ -21,16 +22,22 @@ mod tests;
 #[cfg(test)]
 mod test_data;
 
-use crate::request_handler::main_async;
+use crate::log::*;
+use crate::request_handler::run;
 
-fn main() {
+#[tokio::main]
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
   let _ = ctrlc::set_handler(move || {
     println!();
     std::process::exit(0);
   });
 
-  if let Err(e) = main_async() {
-    eprintln!("Error in main_async: {:?}", e);
-    std::process::exit(1);
+  match run().await {
+    Err(e) => {
+      log_error!("Service handler failed: {}", e);
+      Ok(())
+    },
+
+    _ => Ok(()),
   }
 }
